Rename types.py to structures.py
