+++ maldump/avs/gdata.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import List
 
 from maldump.parsers.gdata_parser import GdataParser
-from maldump.types import Quarantine, QuarEntry
+from maldump.structures import Quarantine, QuarEntry
 
 
 class GData(Quarantine):
