+++ maldump/avs/kaspersky.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import List
 
 from maldump.parsers.kaspersky_parser import KasperskyParser
-from maldump.types import Quarantine, QuarEntry
+from maldump.structures import Quarantine, QuarEntry
 
 
 class Kaspersky(Quarantine):
