+++ maldump/avs/malwarebytes.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import List
 
 from maldump.parsers.malwarebytes_parser import MalwarebytesParser
-from maldump.types import Quarantine, QuarEntry
+from maldump.structures import Quarantine, QuarEntry
 
 
 class Malwarebytes(Quarantine):
