+++ maldump/parsers/malwarebytes_parser.py
@@ -3,7 +3,7 @@ import os.path
 from datetime import datetime as dt
 from typing import List
 
-from maldump.types import QuarEntry
+from maldump.structures import QuarEntry
 from maldump.utils import CustomArc4
